Have travis test documentation build too
rename `requirements.txt` as `test-requirements.txt`; that is the
purpose of this file; normal requirement specification is done properly
in `setup.py`.

Update the travis config to use the new filename and split the script
into two commands; the unittests and use `rst2html.py` to ensure the
docs are compiling correctly.