add count_recordsets function
Similar to count_records and count_media, the new count_recordsets
function returns the number of recordsets in iDigBio.

The new function is added to both the JSON and Pandas client.

count_records has addtional query logic of checking for ingest=true
instead of the bare summary endpoint since this more accurately
reflects the number of recordsets "in" iDigBio.